requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.1
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jsonschema_pydantic==0.6
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-groq==0.3.2
langchain-mcp-adapters==0.1.0
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.37
litellm==1.70.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.0
mcp-use==1.2.13
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
ollama==0.4.8
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pillow==11.2.1
praisonaiagents==0.0.83
propcache==0.3.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.13.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.3
rpds-py==0.25.0
ruff==0.11.8
safetensors==0.5.3
schedule==1.2.2
sentence-transformers==4.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.5
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
